requirements.txt
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4       
click==8.1.7
click-plugins==1.1.1    
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1        
executing==2.0.1        
fiona==1.9.6
Flask==3.0.3
geopandas==1.0.1        
ipykernel==6.29.5       
ipython==8.26.0
itsdangerous==2.2.0     
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2   
jupyter_core==5.7.2     
MarkupSafe==2.1.5       
matplotlib-inline==0.1.7
nest-asyncio==1.6.0     
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
Pygments==2.18.0
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
shapely==2.0.5
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.3